+++ main/adapter/wired/cdi.h
@@ -10,7 +10,7 @@
 void cdi_meta_init(struct generic_ctrl *ctrl_data);
 void cdi_init_buffer(int32_t dev_mode, struct wired_data *wired_data);
 void cdi_from_generic(int32_t dev_mode, struct generic_ctrl *ctrl_data, struct wired_data *wired_data);
-void cdi_kb_id_to_scancode(uint8_t dev_id, uint8_t type, uint8_t id);
+void cdi_kb_id_to_scancode(uint32_t dev_id, uint8_t type, uint8_t id);
 void cdi_gen_turbo_mask(struct wired_data *wired_data);
 
 #endif /* _CDI_H_ */
